Fixinf Routing & move directory
