+++ src/config/Routes.js
@@ -1,12 +1,8 @@
 import RoutesLanding from "../modules/landing_page/LandingRoutes";
 import RoutesHome from "../modules/home/homeRoutes";
-import RouteMataPelajaran from "../modules/mata_pelajaran/mataPelajaranRoutes";
 import Layout from "../components/layout/Layout";
-import RouteKelas from "../modules/kelas/kelasRoutes";
-import RouteSiswa from "../modules/siswa/SiswaRoutes";
-import RouteSchedule from "../modules/schedule/scheduleRoutes";
 import RoutesGuru from "../modules/guru/GuruRoutes";
-import RoutesAbsensi from "../modules/guru/absensi/AbsensiRoutes";
+import RouteTU from "../modules/tu/TURoutes";
 
 const ContainerRoutes = [
   { path: '/school', name: 'layout', component: Layout },
@@ -15,12 +11,8 @@ const ContainerRoutes = [
 
 const MenuRoutes = [
   ...RoutesHome,
-  ...RouteMataPelajaran,
-  ...RouteKelas,
-  ...RouteSiswa,
-  ...RouteSchedule,
+  ...RouteTU,
   ...RoutesGuru,
-  ...RoutesAbsensi,
 ]
 
 export default {
